winewayland.drv: Ensure Vulkan parent surface is mapped with proper size.
The client area subsurface (the target of Vulkan rendering) is not going
to be presented properly (or at all) by the compositor if its parent
surface is not mapped or doesn't have the right size, even though the
window may be visible from Wine's perspective. To avoid this issue,
ensure that the parent surface has up-to-date contents in terms of size,
in case these haven't been provided yet, or will never be provided
(e.g., when the Vulkan rendering is fullscreen).