wined3d: Provide a null counter BO for slots with no corresponding XFB buffer bound.
The Vulkan specification does not currently mandate this. However, an
(unfortunately internal) Khronos discussion determined that this is an
oversight, and that many drivers likely expect that no counter buffer will be
bound if no XFB buffer is bound. Accordingly the specification will probably
mandate this in the future.

lavapipe is such a driver. This fixes a crash with llvmpipe in
test_index_buffer_offset().