wined3d: Set the active texture in sampler_texdim().
This was probably accidentally omitted in 51e64b3fe95afb3246c8690d9ea1e7b4c38ce2b4.

In practice, this is guaranteed to work anyway, since due to the state table
construction the equivalent misc handler, i.e. sampler(), is always guaranteed
to be called right before the fragment pipeline handler for each stage. However,
this is a subtle detail and best not to be depended on.