mshtml/tests: Improve the iframe navigation test.
It was confusing before since it made it seem like it might use the outer
window, while in fact the document is unchanged on native. Now the "new"
doc is used for navigating, since it's already checked to be the same as
the iframes_doc (but that test fails in wine and is todo_wine).