winewayland.drv: Lock the pointer if confined without cursor.
Pointer confinement may only be enabled by the compositor if the pointer
enters the confine region. If the region is small (as in many
mouselook cases) it's very likely that this will never happen and the
pointer will remained unconfined.

To allow mouselook to work more reliably, prefer to lock the pointer if
a window is confined and the cursor is not visible.