winewayland.drv: Apply surface configuration during Vulkan presentation.
When we have a Vulkan client subsurface, and depending on timings and
window state (e.g., fullscreen), we may not get the chance to apply new
surface configurations when presenting content to the main window
surface. To ensure new configurations are applied, also check for and
apply/ack configurations during Vulkan presentation.

This fixes some games which resize when they become fullscreen, but are
not properly placed by the compositor, because they never ack the
fullscreen state.