winemac.drv: Hide app's dock icon when it wouldn't have a taskbar item on Windows.
Hide the icon when an app has no visible windows, or when all of its
visible windows would have no taskbar entry (i.e., they are owned or
have WS_EX_TOOLWINDOW or WS_EX_NOACTIVATE, but not WS_EX_APPWINDOW).
The dock icon returns if those conditions are no longer satisfied.

This behavior is behind a Mac Driver registry key,
EagerDockIconHiding (defaulting to true), to allow toggling it on a
per-app or global basis.