shell32: Always use IContextMenu::InvokeCommand() when selecting an item from the context menu.
This reverts one of the changes included in d30dfd24d6ab2f52bb4f29aee59e75238c648fc4.

Relying on FCIDM_SHVIEW_OPEN snoops into the internals of IContextMenu for no
good reason.

While I haven't tested whether ICommDlgBrowser::OnDefaultCommand() is triggered
in this case, the documentation states that it is "called when a user
double-clicks in the view or presses the ENTER key"; this is neither scenario.

Simply let the context menu call ShellExecuteEx().