winevulkan: Allow only one vulkan surface at a time for an HWND.
Fixes bb872831deb48382e4733979e00243e5122385d0 which lost this from the
winex11 driver. This might be something we want to relax at some point
to allow multiple APIs to draw to the same HWND, but it was done like
that before and it is technically a regression.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=56380