wined3d: Remove no longer used support for drawing to an onscreen render target.
Now that we require ORM_FBO, wined3d backbuffers are always offscreen (and we
never draw directly to the frontbuffer, which is the only truly onscreen
surface).

We could reintroduce support for drawing directly to the backbuffer, and this
would presumably improve performance by avoiding a blit from the wined3d
backbuffer to the real GL backbuffer.

However, this is not always possible or performant due to mismatching
dimensions, or the need to use a depth buffer which is also used with offscreen
RTVs, and for d3d1-9 it may not be possible at all due to issues related to
preserving invariance between flipped and non-flipped geometry.

It *may* be possible to identify situations where we can render directly to the
backbuffer for d3d10+, but at that point we can rely on the Vulkan backend to
work, and given other unfixable performance problems with the GL backend (most
notably the inability to properly stream index buffer data) it does not really
make much sense to go out of our way to perform such a minor optimization there.