winex11: Fix the exposed window surface region combination.
The surface clip region is the region of the window that is visible,
the expose_surface rect is the rect that has been exposed from the host
point of view.

We want to invalidate the part that has been exposed and is visible, so
we need to subtract the complement of the visible region from the rect.