wined3d: Retrieve the VkCommandBuffer from wined3d_context_vk after executing RTV barriers.
Part of beginning a render pass involves executing an RTV barrier, which itself
needs to call wined3d_context_vk_get_command_buffer(). However, that function
may decide to submit the command buffer, in order to prevent resource buildup,
or [in the future] because it has been some length of time since the last
submission.

Therefore we cannot retrieve and store a VkCommandBuffer pointer before
executing an RTV barrier and then use it later.