d3dx9: Fix destination rectangles passed from D3DXLoadSurfaceFromMemory() to d3dx_load_image_from_memory().
d3dx_load_image_from_memory() expects the address of the dst_memory
argument to represent the top left corner of the surface. If a surface
is locked using a rectangle that doesn't start at the top left corner
of the surface, the rectangle we pass to d3dx_load_image_from_memory()
needs to be adjusted so that it properly represents the locked area.

Signed-off-by: Connor McAdams <cmcadams@codeweavers.com>