win32u: Set the virtual desktop display frequency to 60Hz.
Fix a regression from ee0aad5c, which changes the virtual desktop display frequency to that of
the host display while adding modes of 60Hz. So when the host display is not 60Hz, we might get
ChangeDisplaySettings() failures when virtual desktop is on because the target mode is not found
because of the frequency difference.