wined3d: Just check the vertex declaration for point size usage.
Do not check the stream info. In practice, this amounts to a difference if the
usage is included in the vertex declaration but the corresponding stream is not
bound. The just-added tests show that in this case it is correct to use
per-vertex point size with a default size of one, rather than treating this case
as equivalent to if the usage was not specified in the vertex declaration.