wined3d: Just check the vertex declaration for normal usage.
Do not check the stream info.

This would make a difference if the usage was included in the vertex declaration
but the corresponding stream is not bound. However, in that case we supply a
default stream containing zero, which provides the same behaviour as if the
normal usage is not included in the vertex declaration.