wined3d: Just check the vertex declaration for texcoord usage.
Do not check the stream info.

This would make a difference if the usage was included in the vertex declaration
but the corresponding stream is not bound; however, in that case we supply a
default stream containing zero, which provides the same behaviour as if the
usage is not included in the vertex declaration.