wined3d: Always output normalized fog coordinates from the vertex shader.
Historically we used a 0-255 scale in cases where the fog coordinate attribute
was passed through directly, apparently in order to avoid doing float division
on the value passed to glFogCoordfEXT(). This benefit was probably always
dubious, and now is simply unnecessary as we no longer use GL FFP fog nor
immediate mode draws. Get rid of the GLSL equivalent code to scale fog up to
the 0-255 range and then back down to 0-1.