odbc32: Don't load libodbc.so dynamically.
Implementations are guarded by native function pointer checks but all functions
have been present in libodbc for a very long time, so it doesn't appear to be
needed anymore.

The loading code checks the LIB_ODBC_DRIVER_MANAGER environment variable to
allow the user to override the system version of the library. This can be
achieved by other means, if it's still needed.