wineps.drv: Call ResetDCW() to update Devmode in the Unix interface.
Previously, PSDRV_ResetDC() only updated the devmode contained in the
print context, which could cause the devmode in the unix interface to be
inconsistent with the print context's devmode. This can cause problems
later when PostScript is being outputted, as get_device_caps() in the
unix interface would reference this inconsistent devmode. By calling
ResetDCW() in PSDRV_ResetDC(), these two devmodes remain in-sync.