win32u: Simplify the logic for driver messages polling.
As we don't need to make a server request to check for the queue status
anymore, we can stop relying on ProcessEvent return status, and only
call it often enough, before and after checking for messages.

We keep a throttle here though, as now that PeekMessage returns very
quickly when no message is found, calling it in a tight loop triggers
a large number of calls into the drivers instead.