quartz: Properly copy data to render surfaces of planar formats in VMR9.
Before this patch, surface data for planar formats is not copied
correctly when the application uses a custom allocator-presenter and
allocates a surface of different size than the VMR9 source.

This patch adds support for performing this copy correctly when the
source dimensions are less or equal than the rendering surface
dimensions.