widl: Ignore strings in is_embedded_complex().
Arrays declared as strings go via an embedded complex
(e.g. "[string] char s[64];").
This matches the logic used to decide struct or array FC.

This commit does nothing yet, since we pass NULL attributes, but we need to
change that so that other types are treated correctly (in particular,
void pointers with [iid_is].)