dssenh/tests: Remove redundant hash test.
Don't test the hashes to be equal before creating a signature. The
correct hash generation is already tested in the test_hash function.

Also reordered the hash creation and destruction, so that one HCRYPTHASH
variable is enough.

Furthermore removing this check caused the signature verification to
fail. Testing on a Windows 10 machine reveals that this is a wine bug.
Therefore marked the signature verification as todo_wine.

Signed-off-by: Sven Püschel <Sven Püschel@akarisu.de>