dssenh/tests: Fix signLen usage.
Remove the use of signLen to store the input data length. The variable
is used to store the generated signature length. Also using the
ARRAY_SIZE macro instead of the signLen, as in case of the signLen
returning an invalid value, we would compare or sign memory outside of
the arrays (as the failing length test check won't abort the whole
test).

Signed-off-by: Sven Püschel <Sven Püschel@akarisu.de>