dssenh/tests: Separate the public key export from the signhash test.
Separate the exporting of the public key out of the signhash test. As
VerifySignature can also use a privete key, the conversion of the
private key to a public key is removed.

The public key export is kept and moved to a separate test. A
precalculated public key is used to check, that the output format is
correct.

Furthermore change the test_verify_signature -> test_signature as the
test only tested primarely the SignHash function and only used
VerifySignature to check the generated signature is valid.

Signed-off-by: Sven Püschel <Sven Püschel@akarisu.de>