ntdll: Populate the SMBIOS with ARM64 ID register values.
This defines a wine-specific SMBIOS table format ontop of the standard
additional processor information block to pass ID registers to wineboot.
The index of each table within the SMBIOS directly maps to the core it
corresponds to and registers are passed using their numerical IDs to
match Windows' registry format and avoid needing to map between names
and IDs on both ends (the unix side needs to manually encode the msr's
anyway).