cmd: Do not try to handle ERROR_FILE_NOT_FOUND from CreateProcessW().
This was introduced in bff25ffe7991bde75904d34ca733f2178b088316.

The commit aimed to fix command lines like `wine cmd /c "notepad test.txt"`.
Such de-quoting is now handled at an earlier point.

Moreover, we explicitly only enter this path when we find an executable (which
was not the case when the commit was written), so we cannot actually get
ERROR_FILE_NOT_FOUND anymore anyway.