explorer: Use the EnableShell option to show or hide the taskbar.
The restores the option behavior to its original state, where it is off
by default except for the "shell" virtual desktop. It can be forced on
to display a taskbar in every virtual desktop, or off to hide it in the
"shell" desktop too.

The standalone systray window still uses the separate ShowSystray option
which can be enabled when not using virtual desktops, or if EnableShell
is off. When EnableShell is on, the systray area will always be visible
in the taskbar.

Fixes: 118810ad3bac91bfeea81e4408059d48c158e8c5