winemac.drv: Call CGWindowListCreateImageFromArray through a dlsym-obtained pointer.
The macOS Sequoia SDK has "obsoleted" this function: calling it
triggers a build error when targeting macOS Sequoia.
It can still be used though, either by targeting pre-Sequoia or by
getting a pointer to the function through dlsym().

This is intended to be temporary until ScreenCaptureKit.framework or
some other approach can be used.