explorer: Support the NoDesktop registry setting.
Setting NoDesktop to a nonzero value hides all desktop icons on Windows.
(It also disables the context menu on the desktop, but since Wine
doesn't have one, that doesn't matter.) Supporting the same option on
Wine would be useful in controlled environments that have a virtual
desktop but the user is not supposed to interact with it.

Strangely, although HKCU overrides HKLM for NoTrayItemsDisplay on
Windows, the reverse is true of NoDesktop.