winex11: Avoid processing RRNotify events in xrandr14_get_id.
This reverts commit b3eb55227aef198a637f80a2fa0bce74aa0a7190.

Since 278b98f5b21b56abdf9229b12b2a747be6917018 it should be unnecessary to
process the events as the current display mode is written to the registry.

This fixes a random deadlock that happens when a window is updating its state
while another thread changes the current display settings.