d3d11/tests: Remove the workaround for RTVs in test_nv12().
And add a few more tests for valid views.

The assertion that Vulkan drivers don't support NV12 render targets was based on
an incomplete understanding of the (rather complicated) spec. Support for plane
views (the only type allowed in d3d11, as shown by tests added here) is
indicated not through the feature flags for the multi-planar format, but through
the feature flags for the view formats for each plane.