wined3d: Use wined3d_texture_download_from_texture() even if the dst texture map binding is not valid.
If we are writing the entire dst resource—and this is always the case for a
download, since wined3d_texture_download_from_texture() does not support partial
downloads—there is no reason to try to preserve any current map binding the
destination resource may have. We are going to load the resource into its map
binding anyway.

In practice this can happen if the destination's valid location is CLEARED.

With the Vulkan renderer, in that case, we currently load the src texture into
SYSMEM and then perform a CPU copy. This change avoids that.